Added files via upload
Write a program that repeatedly prompts a user for integer numbers until the user enters 'done'.
Once 'done' is entered, print out the largest and smallest of the numbers.
If the user enters anything other than a valid number catch it with a try/except and
output an appropriate message and ignore the input. Enter the numbers from the book for problem.